backend/requirements.txt
# requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
gunicorn

# AI / Computer Vision
ultralytics>=8.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0

# Settings
pydantic-settings==2.2.1
requests==2.31.0
pyyaml==6.0.1
